azure_ml/pyproject.toml
[project]
name = "azure-ml"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "azure-ai-ml>=1.11.0",
    "azure-identity>=1.14.0",
    "azureml-mlflow>=1.53.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
    "keras>=2.13.1",
    "keras-tuner>=1.4.0",
    "matplotlib>=3.7.0",
    "mlflow>=2.8.0",
    "numpy>=1.24.0",
    "opencv-python>=4.8.0",
    "optuna>=3.4.0",
    "pandas>=2.0.0",
    "pillow>=10.0.0",
    "plotly>=5.17.0",
    "pytest>=7.4.0",
    "scikit-learn>=1.3.0",
    "seaborn>=0.12.0",
    "tensorflow[and-cuda]>=2.20.0",
    "tqdm>=4.66.0",
]
